+++ CHANGELOG.md
@@ -60,6 +60,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [#695](https://github.com/alleslabs/celatone-frontend/pull/695) api v1 - contract states
 - [#678](https://github.com/alleslabs/celatone-frontend/pull/678) api v1 - contract table counts
 
+### Bug fixes
+
+- [#707](https://github.com/alleslabs/celatone-frontend/pull/707) contract info api fetching error and module proposal on viewmore
+
 ## v1.4.0
 
 ### Features